Move reboot standby endpoint to failover.reboot service
